Rename Char Case rules as per #106
