+++ src/Locator/SanitizeLocator.php
@@ -37,16 +37,13 @@ class SanitizeLocator extends Locator
             'between'               => function () { return new Sanitize\Between(); },
             'bool'                  => function () { return new Sanitize\Boolean(); },
             'callback'              => function () { return new Sanitize\Callback(); },
-            'caseLower'             => function () { return new Sanitize\CaseLower(); },
-            'caseLowerFirst'        => function () { return new Sanitize\CaseLowerFirst(); },
-            'caseTitle'             => function () { return new Sanitize\CaseTitle(); },
-            'caseUpper'             => function () { return new Sanitize\CaseUpper(); },
-            'caseUpperFirst'        => function () { return new Sanitize\CaseUpperFirst(); },
             'dateTime'              => function () { return new Sanitize\DateTime(); },
             'field'                 => function () { return new Sanitize\Field(); },
             'float'                 => function () { return new Sanitize\Double(); },
             'int'                   => function () { return new Sanitize\Integer(); },
             'isbn'                  => function () { return new Sanitize\Isbn(); },
+            'lowercase'             => function () { return new Sanitize\Lowercase(); },
+            'lowercaseFirst'        => function () { return new Sanitize\LowercaseFirst(); },
             'max'                   => function () { return new Sanitize\Max(); },
             'min'                   => function () { return new Sanitize\Min(); },
             'now'                   => function () { return new Sanitize\Now(); },
@@ -59,7 +56,10 @@ class SanitizeLocator extends Locator
             'strlenBetween'         => function () { return new Sanitize\StrlenBetween(); },
             'strlenMax'             => function () { return new Sanitize\StrlenMax(); },
             'strlenMin'             => function () { return new Sanitize\StrlenMin(); },
+            'titlecase'             => function () { return new Sanitize\Titlecase(); },
             'trim'                  => function () { return new Sanitize\Trim(); },
+            'uppercase'             => function () { return new Sanitize\Uppercase(); },
+            'uppercaseFirst'        => function () { return new Sanitize\UppercaseFirst(); },
             'uuid'                  => function () { return new Sanitize\Uuid(); },
             'uuidHexonly'           => function () { return new Sanitize\UuidHexonly(); },
             'value'                 => function () { return new Sanitize\Value(); },
