+++ src/Locator/ValidateLocator.php
@@ -37,11 +37,6 @@ class ValidateLocator extends Locator
             'between'               => function () { return new Validate\Between(); },
             'bool'                  => function () { return new Validate\Boolean(); },
             'callback'              => function () { return new Validate\Callback(); },
-            'caseLower'             => function () { return new Validate\CaseLower(); },
-            'caseLowerFirst'        => function () { return new Validate\CaseLowerFirst(); },
-            'caseTitle'             => function () { return new Validate\CaseTitle(); },
-            'caseUpper'             => function () { return new Validate\CaseUpper(); },
-            'caseUpperFirst'        => function () { return new Validate\CaseUpperFirst(); },
             'creditCard'            => function () { return new Validate\CreditCard(); },
             'dateTime'              => function () { return new Validate\DateTime(); },
             'email'                 => function () { return new Validate\Email(); },
@@ -49,11 +44,13 @@ class ValidateLocator extends Locator
             'equalToValue'          => function () { return new Validate\EqualToValue(); },
             'float'                 => function () { return new Validate\Double(); },
             'inKeys'                => function () { return new Validate\InKeys(); },
-            'int'                   => function () { return new Validate\Integer(); },
             'inValues'              => function () { return new Validate\InValues(); },
+            'int'                   => function () { return new Validate\Integer(); },
             'ip'                    => function () { return new Validate\Ip(); },
             'isbn'                  => function () { return new Validate\Isbn(); },
             'locale'                => function () { return new Validate\Locale(); },
+            'lowercase'             => function () { return new Validate\Lowercase(); },
+            'lowercaseFirst'        => function () { return new Validate\LowercaseFirst(); },
             'max'                   => function () { return new Validate\Max(); },
             'min'                   => function () { return new Validate\Min(); },
             'regex'                 => function () { return new Validate\Regex(); },
@@ -64,8 +61,11 @@ class ValidateLocator extends Locator
             'strlenBetween'         => function () { return new Validate\StrlenBetween(); },
             'strlenMax'             => function () { return new Validate\StrlenMax(); },
             'strlenMin'             => function () { return new Validate\StrlenMin(); },
+            'titlecase'             => function () { return new Validate\Titlecase(); },
             'trim'                  => function () { return new Validate\Trim(); },
             'upload'                => function () { return new Validate\Upload(); },
+            'uppercase'             => function () { return new Validate\Uppercase(); },
+            'uppercaseFirst'        => function () { return new Validate\UppercaseFirst(); },
             'url'                   => function () { return new Validate\Url(); },
             'uuid'                  => function () { return new Validate\Uuid(); },
             'uuidHexonly'           => function () { return new Validate\UuidHexonly(); },
