+++ src/Rule/Validate/UppercaseFirst.php
@@ -6,28 +6,28 @@
  * @license http://opensource.org/licenses/bsd-license.php BSD
  *
  */
-namespace Aura\Filter\Rule\Sanitize;
+namespace Aura\Filter\Rule\Validate;
 
 use Aura\Filter\Rule\AbstractCharCase;
 
 /**
  *
- * Sanitizes a string to lowercase.
+ * Validates that the string begins with uppercase.
  *
  * @package Aura.Filter
  *
  */
-class CaseLower extends AbstractCharCase
+class UppercaseFirst extends AbstractCharCase
 {
     /**
      *
-     * Sanitizes a string to lowercase.
+     * Validates that the string begins with lowercase.
      *
      * @param object $subject The subject to be filtered.
      *
      * @param string $field The subject field name.
      *
-     * @return bool True if the value was sanitized, false if not.
+     * @return bool True if valid, false if not.
      *
      */
     public function __invoke($subject, $field)
@@ -36,7 +36,7 @@ class CaseLower extends AbstractCharCase
         if (! is_scalar($value)) {
             return false;
         }
-        $subject->$field = $this->strtolower($value);
-        return true;
+
+        return $this->ucfirst($value) == $value;
     }
 }
