--- tests/Rule/Validate/CaseUpperFirstTest.php
@@ -1,28 +1,0 @@
-<?php
-
-namespace Aura\Filter\Rule\Validate;
-
-class CaseUpperFirstTest extends AbstractValidateTest
-{
-    public function providerIs()
-    {
-        return array(
-            array('Ab Cd'),
-            array('EFGH'),
-            array('АБ ВВ'),
-            array('Фг ег'),
-        );
-    }
-
-    public function providerIsNot()
-    {
-        return array(
-            array(array()),
-            array('aBCD'),
-            array('ef GH'),
-            array('аб ВВ'),
-            array('фГ ЕГ'),
-        );
-    }
-}
-
